+++ sdk/storage/azure-storage-file/azure/storage/file/aio/file_service_client_async.py
@@ -26,7 +26,7 @@ from .models import SharePropertiesPaged
 
 if TYPE_CHECKING:
     from datetime import datetime
-    from .._shared.models import ResourceTypes, AccountPermissions
+    from .._shared.models import ResourceTypes, AccountSasPermissions
     from ..models import Metrics, CorsRule, ShareProperties
 
 
@@ -65,7 +65,8 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         account URL already has a SAS token. The value can be a SAS token string or an account
         shared access key.
 
-    Example:
+    .. admonition:: Example:
+
         .. literalinclude:: ../tests/test_file_samples_authentication_async.py
             :start-after: [START create_file_service_client]
             :end-before: [END create_file_service_client]
@@ -99,7 +100,8 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
             The timeout parameter is expressed in seconds.
         :rtype: ~azure.storage.file._generated.models.StorageServiceProperties
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START get_service_properties]
                 :end-before: [END get_service_properties]
@@ -128,21 +130,22 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param hour_metrics:
             The hour metrics settings provide a summary of request
             statistics grouped by API in hourly aggregates for files.
-        :type hour_metrics: ~azure.storage.file.models.Metrics
+        :type hour_metrics: ~azure.storage.file.Metrics
         :param minute_metrics:
             The minute metrics settings provide request statistics
             for each minute for files.
-        :type minute_metrics: ~azure.storage.file.models.Metrics
+        :type minute_metrics: ~azure.storage.file.Metrics
         :param cors:
             You can include up to five CorsRule elements in the
             list. If an empty list is specified, all CORS rules will be deleted,
             and CORS will be disabled for the service.
-        :type cors: list(:class:`~azure.storage.file.models.CorsRule`)
+        :type cors: list(:class:`~azure.storage.file.CorsRule`)
         :param int timeout:
             The timeout parameter is expressed in seconds.
         :rtype: None
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START set_service_properties]
                 :end-before: [END set_service_properties]
@@ -182,9 +185,10 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param int timeout:
             The timeout parameter is expressed in seconds.
         :returns: An iterable (auto-paging) of ShareProperties.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.storage.file.models.ShareProperties]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.storage.file.ShareProperties]
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START fsc_list_shares]
                 :end-before: [END fsc_list_shares]
@@ -229,9 +233,10 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
             Quota in bytes.
         :param int timeout:
             The timeout parameter is expressed in seconds.
-        :rtype: ~azure.storage.file.aio.share_client_async.ShareClient
+        :rtype: ~azure.storage.file.aio.ShareClient
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START fsc_create_shares]
                 :end-before: [END fsc_create_shares]
@@ -258,14 +263,15 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param share_name:
             The share to delete. This can either be the name of the share,
             or an instance of ShareProperties.
-        :type share_name: str or ~azure.storage.file.models.ShareProperties
+        :type share_name: str or ~azure.storage.file.ShareProperties
         :param bool delete_snapshots:
             Indicates if snapshots are to be deleted.
         :param int timeout:
             The timeout parameter is expressed in seconds.
         :rtype: None
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START fsc_delete_shares]
                 :end-before: [END fsc_delete_shares]
@@ -286,13 +292,14 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param share:
             The share. This can either be the name of the share,
             or an instance of ShareProperties.
-        :type share: str or ~azure.storage.file.models.ShareProperties
+        :type share: str or ~azure.storage.file.ShareProperties
         :param str snapshot:
             An optional share snapshot on which to operate.
         :returns: A ShareClient.
-        :rtype: ~azure.storage.file.aio.share_client_async.ShareClient
+        :rtype: ~azure.storage.file.aio.ShareClient
 
-        Example:
+        .. admonition:: Example:
+
             .. literalinclude:: ../tests/test_file_samples_service_async.py
                 :start-after: [START get_share_client]
                 :end-before: [END get_share_client]